pyproject.toml
[project]
name = "acgn-assistant"
version = "0.1.0"
description = "ACGN咨询助手 - ACGN资讯/推荐助手（FastAPI + SQLite）"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.34.0",
  "sqlmodel==0.0.22",
  "psycopg[binary]==3.2.4",
  "pydantic-settings==2.6.1",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "bcrypt==3.2.2",
  "python-multipart==0.0.20",
  "httpx==0.28.1",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
]

[tool.uv]
# uv 可直接读取 pyproject 依赖
